Move timm model wrapper to own file, add improved OpenAI attribution to relevant files
