+++ src/open_clip/tokenizer.py
@@ -1,3 +1,7 @@
+""" CLIP tokenizer
+
+Copied from https://github.com/openai/CLIP. Originally MIT License, Copyright (c) 2021 OpenAI.
+"""
 import gzip
 import html
 import os
